--- a/english/step0.go
+++ b/english/step0.go
@@ -1,16 +1,18 @@
 package english
 
 import (
+	"unicode/utf8"
+
 	"github.com/kljensen/snowball/snowballword"
 )
 
 // Step 0 is to strip off apostrophes and "s".
-//
 func step0(w *snowballword.SnowballWord) bool {
-	suffix, suffixRunes := w.FirstSuffix("'s'", "'s", "'")
+	suffix := w.FirstSuffix("'s'", "'s", "'")
 	if suffix == "" {
 		return false
 	}
-	w.RemoveLastNRunes(len(suffixRunes))
+	sLen := utf8.RuneCountInString(suffix)
+	w.RemoveLastNRunes(sLen)
 	return true
 }