--- a/english/step1a.go
+++ b/english/step1a.go
@@ -1,20 +1,21 @@
 package english
 
 import (
+	"unicode/utf8"
+
 	"github.com/kljensen/snowball/snowballword"
 )
 
 // Step 1a is normalization of various special "s"-endings.
-//
 func step1a(w *snowballword.SnowballWord) bool {
 
-	suffix, suffixRunes := w.FirstSuffix("sses", "ied", "ies", "us", "ss", "s")
+	suffix := w.FirstSuffix("sses", "ied", "ies", "us", "ss", "s")
 	switch suffix {
 
 	case "sses":
 
 		// Replace by ss
-		w.ReplaceSuffixRunes(suffixRunes, []rune("ss"), true)
+		w.ReplaceSuffixRunes([]rune(suffix), []rune("ss"), true)
 		return true
 
 	case "ies", "ied":
@@ -28,7 +29,7 @@ func step1a(w *snowballword.SnowballWord) bool {
 		} else {
 			repl = "ie"
 		}
-		w.ReplaceSuffixRunes(suffixRunes, []rune(repl), true)
+		w.ReplaceSuffixRunes([]rune(suffix), []rune(repl), true)
 		return true
 
 	case "us", "ss":
@@ -37,14 +38,14 @@ func step1a(w *snowballword.SnowballWord) bool {
 		return false
 
 	case "s":
-
 		// Delete if the preceding word part contains a vowel
 		// not immediately before the s (so gas and this retain
 		// the s, gaps and kiwis lose it)
 		//
+		sLen := utf8.RuneCountInString(suffix)
 		for i := 0; i < len(w.RS)-2; i++ {
 			if isLowerVowel(w.RS[i]) {
-				w.RemoveLastNRunes(len(suffixRunes))
+				w.RemoveLastNRunes(sLen)
 				return true
 			}
 		}