--- a/english/step1b.go
+++ b/english/step1b.go
@@ -1,14 +1,16 @@
 package english
 
 import (
+	"unicode/utf8"
+
 	"github.com/kljensen/snowball/snowballword"
 )
 
 // Step 1b is the normalization of various "ly" and "ed" sufficies.
-//
 func step1b(w *snowballword.SnowballWord) bool {
 
-	suffix, suffixRunes := w.FirstSuffix("eedly", "ingly", "edly", "ing", "eed", "ed")
+	suffix := w.FirstSuffix("eedly", "ingly", "edly", "ing", "eed", "ed")
+	sLen := utf8.RuneCountInString(suffix)
 
 	switch suffix {
 
@@ -19,14 +21,14 @@ func step1b(w *snowballword.SnowballWord) bool {
 	case "eed", "eedly":
 
 		// Replace by ee if in R1
-		if len(suffixRunes) <= len(w.RS)-w.R1start {
-			w.ReplaceSuffixRunes(suffixRunes, []rune("ee"), true)
+		if sLen <= len(w.RS)-w.R1start {
+			w.ReplaceSuffixRunes([]rune(suffix), []rune("ee"), true)
 		}
 		return true
 
 	case "ed", "edly", "ing", "ingly":
 		hasLowerVowel := false
-		for i := 0; i < len(w.RS)-len(suffixRunes); i++ {
+		for i := 0; i < len(w.RS)-sLen; i++ {
 			if isLowerVowel(w.RS[i]) {
 				hasLowerVowel = true
 				break
@@ -45,11 +47,11 @@ func step1b(w *snowballword.SnowballWord) bool {
 			originalR2start := w.R2start
 
 			// Delete if the preceding word part contains a vowel
-			w.RemoveLastNRunes(len(suffixRunes))
+			w.RemoveLastNRunes(sLen)
 
 			// ...and after the deletion...
 
-			newSuffix, newSuffixRunes := w.FirstSuffix("at", "bl", "iz", "bb", "dd", "ff", "gg", "mm", "nn", "pp", "rr", "tt")
+			newSuffix := w.FirstSuffix("at", "bl", "iz", "bb", "dd", "ff", "gg", "mm", "nn", "pp", "rr", "tt")
 			switch newSuffix {
 
 			case "":
@@ -68,7 +70,7 @@ func step1b(w *snowballword.SnowballWord) bool {
 			case "at", "bl", "iz":
 
 				// If the word ends "at", "bl" or "iz" add "e"
-				w.ReplaceSuffixRunes(newSuffixRunes, []rune(newSuffix+"e"), true)
+				w.ReplaceSuffixRunes([]rune(newSuffix), []rune(newSuffix+"e"), true)
 
 			case "bb", "dd", "ff", "gg", "mm", "nn", "pp", "rr", "tt":