--- a/english/step4.go
+++ b/english/step4.go
@@ -1,6 +1,8 @@
 package english
 
 import (
+	"unicode/utf8"
+
 	"github.com/kljensen/snowball/snowballword"
 )
 
@@ -17,14 +19,15 @@ import (
 func step4(w *snowballword.SnowballWord) bool {
 
 	// Find all endings in R1
-	suffix, suffixRunes := w.FirstSuffix(
+	suffix := w.FirstSuffix(
 		"ement", "ance", "ence", "able", "ible", "ment",
 		"ent", "ant", "ism", "ate", "iti", "ous", "ive",
 		"ize", "ion", "al", "er", "ic",
 	)
+	sLen := utf8.RuneCountInString(suffix)
 
 	// If it does not fit in R2, do nothing.
-	if len(suffixRunes) > len(w.RS)-w.R2start {
+	if sLen > len(w.RS)-w.R2start {
 		return false
 	}
 
@@ -40,7 +43,7 @@ func step4(w *snowballword.SnowballWord) bool {
 		if rsLen >= 4 {
 			switch w.RS[rsLen-4] {
 			case 115, 116:
-				w.RemoveLastNRunes(len(suffixRunes))
+				w.RemoveLastNRunes(sLen)
 				return true
 			}
 
@@ -49,7 +52,7 @@ func step4(w *snowballword.SnowballWord) bool {
 	}
 
 	// Handle basic replacements
-	w.RemoveLastNRunes(len(suffixRunes))
+	w.RemoveLastNRunes(sLen)
 	return true
 
 }