… .github/workflows/test.yml
… README.md
… english/english_test.go
… english_vocab/vocab_test.go
… french/step2b.go
… french/step4.go
… french/step5.go
… norwegian/step1.go
… norwegian/step2.go
… norwegian/step3.go
… russian/step1.go
… russian/step2.go
… russian/step3.go
… russian/step4.go
… snowballword/snowballword.go
… snowballword/snowballword_test.go
… spanish/step0.go
… spanish/step1.go
… spanish/step2a.go
… spanish/step2b.go
… spanish/step3.go
… swedish/step1.go
… swedish/step2.go
… swedish/step3.go