Remove second return parameter `suffixRunes` from many functions, as it’s not used by every caller, and requires an allocation. In many cases, just its length is needed, so use `utf8.RuneCountInString`. In cases where `suffixRunes` is needed, create it locally.

Before:

```
Benchmark_Stem-8   	      50950417 ns/op	   4.41 MB/s	38536284 B/op	 2066112 allocs/op
```

After:

```
Benchmark_Stem-8   	      25477538 ns/op	   8.82 MB/s	 1377490 B/op	   58406 allocs/op
```

~2x throughput improvement and ~30x allocation reduction.

(nit: my formatter seems to have removed a bunch of blank `//` lines; not deliberate by me, let me know if it matters.)