@clipperhouse 2024-08-13
Great! I suspect the failures above have to do with old Go versions. I updated go.mod and test.yml for Go 1.19 and above (and got rid of the cache), let’s see if that fixes it.

(If you’re not ready to update Go version, that’s fine, but at least we’ll have some troubleshooting.)