@kljensen 2024-08-13
Thanks @clipperhouse. Three quick questions. (I haven't looked at this code in a while, so apologies if these are non-sensical.)

1. It seems like there were a lot of white space changes. Can you help me understand those?
2. How do you feel about using `suffixLength` or `suffixLen` instead of `sLen` throughout?
3. I like your benchmark above. Is there a way to incorporate that into the automated tests?