@clipperhouse 2024-08-13
Cheers @kljensen 

1) I didn’t do any intentional whitespace changes, but the Go formatter seemed to. In particular it removed a lot of empty `//` comment lines. [Here’s some explanation](https://www.perplexity.ai/search/in-golang-with-vs-code-and-gop-yEvFndSLTMasBuPXJ9GA6A). On GitHub, you can “Hide whitespace changes” in the PR, under the little settings wheel icon at the top. I can go through and try to undo those formats if it’s important.
2) Done!
3) We could change the test.yml action to `go test ./... -benchmem -bench=.`, which will run (log) the benchmarks in addition to the tests. In terms of doing anything with benchmark results, like tracking over time, I don’t know a good way to do that. I’m sure something exists.