@clipperhouse 2024-08-13
Done (I assume you mean contributors in the README?) and thanks! I usually squash PRs as well.